[REEF-237] Add a DriverStarted handler that generates the ClassHierarchy
This adds `ClassHierarchyGeneratingDriverStartObserver` and registers it
as a `DriverStartHandler`. This new class generates the class hierarchy
for the assemblies found in `reef/global`.

HelloREEF is adapted to no longer generate the class hiearchy.

JIRA:
  [REEF-237](https://issues.apache.org/jira/browse/REEF-237)

Pull Request:
  This closes #
